Android Sync Script, Default Settings, Bugfixes
Biggest news is I uploaded the script I am using for my nightly
rsync of my podcasts over WiFi to an Android phone. I'm using SSH,
rsync, and openssh key authentication. The script scans the podcast
folder on the phone for files older than 3 days and deletes them.
Then, a temporary folder is built containing symlinks of the files
to be sent to the phone. I do this because my podcasts are stored in
some folders, but on my phone I prefer to just keep them all together
in a main folder. Next the script rsyncs the folder over and applies 777
permissions to all files there. Finally, the phone is rebooted. This is needed to forced media players to
"rescan" the library.

I realized that in the last commit I left in the settings I was using,
so I restored them to the "normal" defaults on this commit.
Fixed some non-formatted time stamps I missed on the last commit.